# GraphWebApi/wwwroot/OpenApi.yaml
openapi: 3.0.0
info: 
  title: "Graph Explorer API" 
  version: "1.0.0" 
servers:
  - url: https://graphexplorerapi.azurewebsites.net/
  - url: https://localhost:44399/
paths:
...
  /samples:
    get:
      tags:
        - samples
      description: Get all sample queries or filter by category, humanName or tip. 
      parameters:
        - name: search   # /samples?search=user
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            'application/json': 
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/sample'
        204:
          description: No content
        404:
          description: Not found
        500:
          description: Internal server error
    post:
      tags:
        - samples
      description: Create a new sample.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/sample'
            example:
              {
                "category": "Security",
                "method": "GET",
                "humanName": "alerts filter by destination address",
                "requestUrl": "/v1.0/security/alerts?$filter=networkConnections/any(d:d/destinationAddress eq '{destination-address}')",
                "docLink": "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/alert_list",
                "tip": "This query requires a destination address. Run https://graph.microsoft.com/v1.0/security/alerts?$top=1 and search the results for a destinationAddress property.",
                "skipTest": false
              }
      responses:
        201: 
          description: Returns created sample with provisioned Id value.
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/sample'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        500:
          description: Internal server error
  /samples/{id}:
    get:
      tags:
        - samples
      description: Get a specific sample query by its Id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/sample'
        204:
          description: No content
        404:
          description: Not found
        500:
          description: Internal server error
    put:
      tags:
        - samples
      description: Update a specific sample query by its Id.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/sample'
            example:
              {
                "category": "Security",
                "method": "GET",
                "humanName": "alerts filter by destination address",
                "requestUrl": "/v1.0/security/alerts?$filter=networkConnections/any(d:d/destinationAddress eq '{destination-address}')",
                "docLink": "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/alert_list",
                "tip": "This query requires a destination address. Run https://graph.microsoft.com/v1.0/security/alerts?$top=1 and search the results for a destinationAddress property.",
                "skipTest": false
              }
      responses:
        200:
          description: Ok
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/sample'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
    delete:
      tags:
        - samples
      description: Delete a specific sample query by its Id.
      security:
        - OAuth2Implicit: [api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Deleted successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: Not found
        500:
          description: Internal server error
...
components:
  schemas:
    sample:
      type: object
      required:
        - category
        - method
        - humanName
        - requestUrl
      properties:
        id:
          type: string
        category:
          type: string
        method:
          type: string
          enum:
            - GET
            - PUT
            - POST
            - DELETE
            - PATCH
        humanName:
          type: string
        requestUrl:
          type: string
        docLink:
          type: string
        tip:
          type: string
        skipTest: 
          type: boolean
        postBody:
          type: string
        headers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
...
  securitySchemes:
    OAuth2Implicit:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow to authenticate with Azure AD.
      flows:
        implicit:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          scopes: 
            api://xxxx-xxxx-xxxx-xxxx-xxxx/SampleQueries.ReadWriteAll: writes, updates and deletes workload samples.